+++ scout/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 try:
     from importlib.metadata import version
-except ImportError:
+except ImportError:  # Backport support for importlib metadata on Python 3.7
     from importlib_metadata import version
 
 __version__ = version("scout-browser")