+++ CHANGELOG.md
@@ -22,6 +22,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Fixed a couple of creations where now was called twice, so updated_at and created_at could differ
 - Deprecated Ubuntu version 18.04 in one GitHub action
 - Panels that have been removed (hidden) should not be visible in views where overlapping gene panels for genes are shown
+- Gene panel test pointing to the right function
 
 ## [4.64]
 ### Added