+++ CHANGELOG.md
@@ -17,13 +17,13 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Show comment badges in related comments tables on general report
 - Updated version of several GitHub actions
 - Migrate from deprecated `pkg_resources` lib to `importlib_resources`
-- Removed duplicated test
 ### Fixed
 - Fixed Sanger order / Cancel order modal close buttons
 - Visibility of SV type in ClinVar submission form
 - Fixed a couple of creations where now was called twice, so updated_at and created_at could differ
 - Deprecated Ubuntu version 18.04 in one GitHub action
 - Panels that have been removed (hidden) should not be visible in views where overlapping gene panels for genes are shown
+- Gene panel test pointing to the right function
 
 ## [4.64]
 ### Added