+++ CHANGELOG.md
@@ -10,6 +10,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Docs on how to create/update the PanelApp green genes as a system admin
 - `individual_id`-parameter to both Gens links
 - Download a gene panel in CSV format from gene panel page
+- Panel gene comments: genes in panels can have comments that describe the gene in a panel context
 ### Changed
 - Always show each case category on caseS page, even if 0 cases in total or after current query
 - Improved sorting of ClinVar submissions
@@ -17,6 +18,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Show comment badges in related comments tables on general report
 - Updated version of several GitHub actions
 - Migrate from deprecated `pkg_resources` lib to `importlib_resources`
+- Dismiss bar on variantS pages is thinner.
+- Dismiss bar on variantS pages can be toggled open or closed for the duration of a login session.
 ### Fixed
 - Fixed Sanger order / Cancel order modal close buttons
 - Visibility of SV type in ClinVar submission form