+++ CHANGELOG.md
@@ -8,6 +8,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Fixed
 - Visibility of `Gene(s)` badges on SV VariantS page
 - Hide dismiss bar on SV page not working well
+- Delivery report PDF download
+- Saving Pipeline version file when loading a case
 
 ## [4.65]
 ### Added